requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
click==7.0                # via pip-tools
django-crispy-forms==1.7.2
django-extensions==2.1.6
django==2.2
jinja2==2.10.1
markupsafe==1.1.1         # via jinja2
pillow==6.0.0
pip-tools==3.6.0
psycopg2==2.8.2
pytz==2019.1              # via django
six==1.12.0               # via django-extensions, pip-tools
sqlparse==0.3.0           # via django
